fix: harden function dryrun lifecycle
