+++ packages/server/lib/controllers/functions/dryrun/postDryrun.ts
@@ -1,159 +1,21 @@
-import db from '@nangohq/database';
 import {
     RemoteFunctionError,
     createFunctionDryrun,
-    getFunctionDryrun as getStoredFunctionDryrun,
-    getFunctionDryrunRow,
     getRemoteFunctionNangoHost,
-    invokeDryrun,
     markFunctionDryrunFailed,
     markFunctionDryrunRunning,
-    markFunctionDryrunSuccess,
-    parseDryrunSuccessOutput,
     prepareAsyncDryrun,
-    remoteFunctionDryrunSandboxTimeoutMs,
-    sandboxApiKeyService,
     toFunctionDryrunCreate
 } from '@nangohq/sandbox';
 import { configService, connectionService } from '@nangohq/shared';
-import { requireEmptyQuery, stringifyError, zodErrorToHTTP } from '@nangohq/utils';
+import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
 import { asyncWrapper } from '../../../utils/asyncWrapper.js';
 import { sendStepError } from '../errors.js';
-import { functionDryrunBodySchema, functionDryrunParamsSchema, functionDryrunResultBodySchema, remoteFunctionDryrunBodySchema } from '../validation.js';
+import { functionDryrunBodySchema } from '../validation.js';
+import { createDryrunSandboxApiKey, defaultFunctionName, requireCustomerKeyId, toFunctionDryrunError } from './helpers.js';
 
-import type { RequestLocals } from '../../../utils/express.js';
-import type { FunctionErrorCode, GetFunctionDryrun, PostFunctionDryrun, PostFunctionDryrunResult, PostRemoteFunctionDryrun } from '@nangohq/types';
-import type { Response } from 'express';
-
-const defaultFunctionName = 'function';
-
-const sandboxApiKeyTimeoutBufferMs = 60 * 1000;
-
-const functionErrorCodes = new Set<string>([
-    'invalid_request',
-    'integration_not_found',
-    'compilation_error',
-    'dryrun_error',
-    'deployment_error',
-    'connection_not_found',
-    'dryrun_not_found',
-    'function_disabled',
-    'execution_environment_unavailable',
-    'timeout',
-    'validation_error'
-] satisfies FunctionErrorCode[]);
-
-type RemoteDryrunResponse = Response<PostRemoteFunctionDryrun['Reply'], Required<RequestLocals>>;
-type FunctionDryrunResponse = Response<PostFunctionDryrun['Reply'], Required<RequestLocals>>;
-type FunctionDryrunResultResponse = Response<PostFunctionDryrunResult['Reply'], Required<RequestLocals>>;
-
-async function createDryrunSandboxApiKey(parentApiKeyId: number, environmentId: number, dryrunId?: string) {
-    return await sandboxApiKeyService.createSandboxApiKey(db.knex, {
-        parentApiKeyId,
-        environmentId,
-        purpose: 'dryrun',
-        ...(dryrunId ? { dryrunId } : {}),
-        expiresAt: new Date(Date.now() + remoteFunctionDryrunSandboxTimeoutMs + sandboxApiKeyTimeoutBufferMs)
-    });
-}
-
-function getRemoteParentCustomerKeyId(res: RemoteDryrunResponse): number | null {
-    if (res.locals['apiKeyAuthSource'] !== 'customer_key' || !res.locals['apiKeyId']) {
-        res.status(403).send({ error: { code: 'forbidden', message: 'Sandbox tokens can only be issued from a customer API key' } });
-        return null;
-    }
-
-    return res.locals['apiKeyId'];
-}
-
-function getFunctionParentCustomerKeyId(res: FunctionDryrunResponse): number | null {
-    if (res.locals['apiKeyAuthSource'] !== 'customer_key' || !res.locals['apiKeyId']) {
-        res.status(403).send({ error: { code: 'forbidden', message: 'Function dryruns can only be started from a customer API key' } });
-        return null;
-    }
-
-    return res.locals['apiKeyId'];
-}
-
-function verifyDryrunResultSandboxToken(res: FunctionDryrunResultResponse, dryrunId: string): boolean {
-    if (res.locals['sandboxTokenPurpose'] !== 'dryrun' || res.locals['sandboxTokenDryrunId'] !== dryrunId) {
-        res.status(403).send({ error: { code: 'forbidden', message: 'This sandbox token is not authorized for this dryrun' } });
-        return false;
-    }
-
-    return true;
-}
-
-export const postRemoteFunctionDryrun = asyncWrapper<PostRemoteFunctionDryrun>(async (req, res) => {
-    const emptyQuery = requireEmptyQuery(req);
-    if (emptyQuery) {
-        res.status(400).send({ error: { code: 'invalid_query_params', errors: zodErrorToHTTP(emptyQuery.error) } });
-        return;
-    }
-
-    const valBody = remoteFunctionDryrunBodySchema.safeParse(req.body);
-    if (!valBody.success) {
-        res.status(400).send({ error: { code: 'invalid_body', errors: zodErrorToHTTP(valBody.error) } });
-        return;
-    }
-
-    const body = valBody.data;
-    const { environment } = res.locals;
-
-    const connectionResult = await connectionService.getConnection(body.connection_id, body.integration_id, environment.id);
-    if (!connectionResult.success || !connectionResult.response) {
-        sendStepError({
-            res,
-            status: 404,
-            error: { type: 'connection_not_found', message: `Connection '${body.connection_id}' was not found for integration '${body.integration_id}'` }
-        });
-        return;
-    }
-
-    const parentCustomerKeyId = getRemoteParentCustomerKeyId(res);
-    if (!parentCustomerKeyId) {
-        return;
-    }
-
-    const sandboxApiKey = await createDryrunSandboxApiKey(parentCustomerKeyId, environment.id);
-    if (sandboxApiKey.isErr()) {
-        sendStepError({ res, status: 500, error: sandboxApiKey.error });
-        return;
-    }
-
-    const startedAt = new Date();
-
-    try {
-        const result = await invokeDryrun({
-            integration_id: body.integration_id,
-            function_name: body.function_name,
-            function_type: body.function_type,
-            code: body.code,
-            environment_name: environment.name,
-            connection_id: body.connection_id,
-            nango_secret_key: sandboxApiKey.value,
-            nango_host: getRemoteFunctionNangoHost(),
-            ...(body.input !== undefined ? { input: body.input } : {}),
-            ...(body.metadata ? { metadata: body.metadata } : {}),
-            ...(body.checkpoint ? { checkpoint: body.checkpoint } : {}),
-            ...(body.last_sync_date ? { last_sync_date: body.last_sync_date } : {})
-        });
-
-        const durationMs = Date.now() - startedAt.getTime();
-        const dryrunOutput = parseDryrunSuccessOutput(result.output);
-
-        res.status(200).send({
-            integration_id: body.integration_id,
-            function_name: body.function_name,
-            function_type: body.function_type,
-            duration_ms: durationMs,
-            ...(dryrunOutput.hasResult ? { result: dryrunOutput.result } : {})
-        });
-    } catch (err) {
-        sendStepError({ res, error: err, ...(err instanceof RemoteFunctionError ? {} : { status: 500 }) });
-    }
-});
+import type { PostFunctionDryrun } from '@nangohq/types';
 
 export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, res) => {
     const emptyQuery = requireEmptyQuery(req);
@@ -170,7 +32,7 @@ export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, r
 
     const body = valBody.data;
     const { environment } = res.locals;
-    const parentCustomerKeyId = getFunctionParentCustomerKeyId(res);
+    const parentCustomerKeyId = requireCustomerKeyId(res, 'Function dryruns can only be started from a customer API key');
     if (!parentCustomerKeyId) {
         return;
     }
@@ -191,7 +53,7 @@ export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, r
         return;
     }
 
-    const dryrun = await createFunctionDryrun({
+    const dryrunResult = await createFunctionDryrun({
         environmentId: environment.id,
         request: {
             integration_id: body.integration_id,
@@ -205,7 +67,12 @@ export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, r
             ...(body.last_sync_date ? { last_sync_date: body.last_sync_date } : {})
         }
     });
+    if (dryrunResult.isErr()) {
+        sendStepError({ res, status: 500, error: dryrunResult.error });
+        return;
+    }
 
+    const dryrun = dryrunResult.value;
     let prepared: Awaited<ReturnType<typeof prepareAsyncDryrun>> | null = null;
     try {
         const sandboxApiKey = await createDryrunSandboxApiKey(parentCustomerKeyId, environment.id, dryrun.id);
@@ -247,7 +114,9 @@ export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, r
 
         res.status(202).send(toFunctionDryrunCreate(running));
     } catch (err) {
-        await prepared?.kill().catch(() => {});
+        await prepared?.kill().catch(() => {
+            // Still mark the dryrun as failed if sandbox cleanup fails.
+        });
         await markFunctionDryrunFailed({
             environmentId: environment.id,
             id: dryrun.id,
@@ -256,112 +125,3 @@ export const postFunctionDryrun = asyncWrapper<PostFunctionDryrun>(async (req, r
         sendStepError({ res, error: err, ...(err instanceof RemoteFunctionError ? {} : { status: 500 }) });
     }
 });
-
-export const getFunctionDryrun = asyncWrapper<GetFunctionDryrun>(async (req, res) => {
-    const emptyQuery = requireEmptyQuery(req);
-    if (emptyQuery) {
-        res.status(400).send({ error: { code: 'invalid_query_params', errors: zodErrorToHTTP(emptyQuery.error) } });
-        return;
-    }
-
-    const valParams = functionDryrunParamsSchema.safeParse(req.params);
-    if (!valParams.success) {
-        res.status(400).send({ error: { code: 'invalid_uri_params', errors: zodErrorToHTTP(valParams.error) } });
-        return;
-    }
-
-    const { environment } = res.locals;
-    const dryrun = await getStoredFunctionDryrun({ environmentId: environment.id, id: valParams.data.id });
-    if (!dryrun) {
-        res.status(404).send({ error: { code: 'dryrun_not_found', message: `Dryrun '${valParams.data.id}' was not found` } });
-        return;
-    }
-
-    res.status(200).send(dryrun);
-});
-
-export const postFunctionDryrunResult = asyncWrapper<PostFunctionDryrunResult>(async (req, res) => {
-    const emptyQuery = requireEmptyQuery(req);
-    if (emptyQuery) {
-        res.status(400).send({ error: { code: 'invalid_query_params', errors: zodErrorToHTTP(emptyQuery.error) } });
-        return;
-    }
-
-    const valParams = functionDryrunParamsSchema.safeParse(req.params);
-    if (!valParams.success) {
-        res.status(400).send({ error: { code: 'invalid_uri_params', errors: zodErrorToHTTP(valParams.error) } });
-        return;
-    }
-
-    if (!verifyDryrunResultSandboxToken(res, valParams.data.id)) {
-        return;
-    }
-
-    const valBody = functionDryrunResultBodySchema.safeParse(req.body);
-    if (!valBody.success) {
-        res.status(400).send({ error: { code: 'invalid_body', errors: zodErrorToHTTP(valBody.error) } });
-        return;
-    }
-
-    const { environment } = res.locals;
-    const current = await getFunctionDryrunRow({ environmentId: environment.id, id: valParams.data.id });
-    if (!current) {
-        res.status(404).send({ error: { code: 'dryrun_not_found', message: `Dryrun '${valParams.data.id}' was not found` } });
-        return;
-    }
-
-    if (current.status === 'success' || current.status === 'failed') {
-        res.status(200).send({ ok: true });
-        return;
-    }
-
-    const body = valBody.data;
-    if (body.status === 'success') {
-        const output = parseDryrunSuccessOutput(body.output);
-        await markFunctionDryrunSuccess({
-            environmentId: environment.id,
-            id: current.id,
-            output: output.output,
-            result: output.result,
-            hasResult: output.hasResult,
-            durationMs: body.duration_ms
-        });
-    } else {
-        await markFunctionDryrunFailed({
-            environmentId: environment.id,
-            id: current.id,
-            output: body.output,
-            durationMs: body.duration_ms,
-            error: {
-                code: normalizeFunctionErrorCode(body.error.code),
-                message: body.error.message,
-                ...(body.error.payload !== undefined ? { payload: body.error.payload } : {})
-            }
-        });
-    }
-
-    res.status(200).send({ ok: true });
-});
-
-function normalizeFunctionErrorCode(code: string | undefined): FunctionErrorCode {
-    if (code && functionErrorCodes.has(code)) {
-        return code as FunctionErrorCode;
-    }
-
-    return 'dryrun_error';
-}
-
-function toFunctionDryrunError(err: unknown): { code: FunctionErrorCode; message: string; payload?: unknown } {
-    if (err instanceof RemoteFunctionError) {
-        return {
-            code: err.code,
-            message: err.message,
-            ...(err.payload !== undefined ? { payload: err.payload } : {})
-        };
-    }
-
-    return {
-        code: 'dryrun_error',
-        message: stringifyError(err)
-    };
-}
