+++ packages/server/lib/routes.public.ts
@@ -36,7 +36,10 @@ import connectionController from './controllers/connection.controller.js';
 import { getPublicEnvironmentVariables } from './controllers/environment/getVariables.js';
 import { postFunctionCompile, postRemoteFunctionCompile } from './controllers/functions/compile/postCompile.js';
 import { postFunctionDeployment, postRemoteFunctionDeploy } from './controllers/functions/deploy/postDeploy.js';
-import { getFunctionDryrun, postFunctionDryrun, postFunctionDryrunResult, postRemoteFunctionDryrun } from './controllers/functions/dryrun/postDryrun.js';
+import { getFunctionDryrun } from './controllers/functions/dryrun/getDryrun.js';
+import { postFunctionDryrun } from './controllers/functions/dryrun/postDryrun.js';
+import { postFunctionDryrunResult } from './controllers/functions/dryrun/postDryrunResult.js';
+import { postRemoteFunctionDryrun } from './controllers/functions/dryrun/postRemoteDryrun.js';
 import { getPublicListIntegrations } from './controllers/integrations/getListIntegrations.js';
 import { postPublicIntegration, postPublicQuickstartIntegration } from './controllers/integrations/postIntegration.js';
 import { deletePublicIntegration } from './controllers/integrations/uniqueKey/deleteIntegration.js';
