+++ packages/shared/lib/utils/utils.ts
@@ -6,7 +6,7 @@ import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc.js';
 import get from 'lodash-es/get.js';
 
-import { isEnterprise, localhostUrl } from '@nangohq/utils';
+import { cloudHost, isCloud, isStaging, localhostUrl, stagingHost } from '@nangohq/utils';
 
 import type { DBConnection, Provider } from '@nangohq/types';
 
@@ -51,7 +51,11 @@ export function getJobsUrl() {
 }
 
 function getServerHost() {
-    return process.env['SERVER_HOST'] || process.env['SERVER_RUN_MODE'] === 'DOCKERIZED' ? 'http://nango-server' : 'http://localhost';
+    if (process.env['SERVER_HOST']) {
+        return process.env['SERVER_HOST'];
+    }
+
+    return process.env['SERVER_RUN_MODE'] === 'DOCKERIZED' ? 'http://nango-server' : 'http://localhost';
 }
 
 export function getServerBaseUrl() {
@@ -165,8 +169,12 @@ export function getLocalOAuthCallbackUrlBaseUrl() {
 }
 
 export function getApiUrl() {
-    if (isEnterprise) {
-        return process.env['NANGO_SERVER_URL'] as string;
+    if (isCloud) {
+        return isStaging ? stagingHost : cloudHost;
+    }
+
+    if (process.env['NANGO_SERVER_URL']) {
+        return process.env['NANGO_SERVER_URL'];
     }
     if (process.env['SERVER_SERVICE_URL']) {
         return process.env['SERVER_SERVICE_URL'];
